+++ CHANGELOG.md
@@ -122,8 +122,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - Added mechanism to extend block header. Block now contains
   key-value storage `additional_headers` which can contain binary data. (#1602)
   
-- Added method `keys_unchecked` to `Group` index which return all the keys from the group.
-(#1614)
+- Added method `keys_unchecked` to `Group` index which return all the keys
+from the group. (#1614)
 
 - `TemporaryDB` can now be cleared. This will remove contents of all indexes
   and erase index metadata. (#1630)